+++ lib/src/bones_api_entity_db_object_directory.dart
@@ -16,8 +16,9 @@ import 'bones_api_entity_db.dart';
 import 'bones_api_entity_db_object.dart';
 import 'bones_api_entity_reference.dart';
 import 'bones_api_extension.dart';
+import 'bones_api_logging.dart';
 
-final _log = logging.Logger('DBObjectDirectoryAdapter');
+final _log = logging.Logger('DBObjectDirectoryAdapter')..registerAsDbLogger();
 
 class DBObjectDirectoryAdapterContext
     implements Comparable<DBObjectDirectoryAdapterContext> {
@@ -48,6 +49,8 @@ class DBObjectDirectoryAdapter
   static void boot() {
     if (_boot) return;
     _boot = true;
+
+    DBObjectAdapter.boot();
 
     DBObjectAdapter.registerAdapter([
       'object.directory',
@@ -683,9 +686,11 @@ class DBObjectDirectoryAdapter
     return entityJsonNormalized;
   }
 
-  Object resolveError(Object error, StackTrace stackTrace) =>
+  Object resolveError(Object error, StackTrace stackTrace, Object? operation) =>
       DBObjectDirectoryAdapterException('error', '$error',
-          parentError: error, parentStackTrace: stackTrace);
+          parentError: error,
+          parentStackTrace: stackTrace,
+          operation: operation);
 
   FutureOr<R> _finishOperation<T, R>(
       TransactionOperation op, T res, PreFinishDBOperation<T, R>? preFinish) {
@@ -706,6 +711,7 @@ class DBObjectDirectoryAdapter
                 e,
                 s,
                 errorResolver: resolveError,
+                operation: op,
                 debugInfo: () => op.toString(),
               ));
     }
@@ -714,6 +720,7 @@ class DBObjectDirectoryAdapter
       return transaction.addExecution<R, DBObjectDirectoryAdapterContext>(
         (c) => f(c),
         errorResolver: resolveError,
+        operation: op,
         debugInfo: () => op.toString(),
       );
     }
@@ -732,6 +739,7 @@ class DBObjectDirectoryAdapter
       return transaction.addExecution<R, DBObjectDirectoryAdapterContext>(
         (c) => f(c),
         errorResolver: resolveError,
+        operation: op,
         debugInfo: () => op.toString(),
       );
     });
@@ -744,7 +752,6 @@ class DBObjectDirectoryAdapterException extends DBObjectAdapterException {
   String get runtimeTypeNameSafe => 'DBObjectDirectoryAdapterException';
 
   DBObjectDirectoryAdapterException(String type, String message,
-      {Object? parentError, StackTrace? parentStackTrace})
-      : super(type, message,
-            parentError: parentError, parentStackTrace: parentStackTrace);
+      {super.parentError, super.parentStackTrace, super.operation})
+      : super(type, message);
 }