+++ lib/src/bones_api_entity_db_postgres.dart
@@ -12,12 +12,13 @@ import 'bones_api_entity_db.dart';
 import 'bones_api_entity_db_sql.dart';
 import 'bones_api_extension.dart';
 import 'bones_api_initializable.dart';
+import 'bones_api_logging.dart';
 import 'bones_api_sql_builder.dart';
 import 'bones_api_types.dart';
 import 'bones_api_utils.dart';
 import 'bones_api_utils_timedmap.dart';
 
-final _log = logging.Logger('DBPostgreSQLAdapter');
+final _log = logging.Logger('DBPostgreSQLAdapter')..registerAsDbLogger();
 
 /// A PostgreSQL adapter.
 class DBPostgreSQLAdapter extends DBSQLAdapter<PostgreSQLExecutionContext>
@@ -30,6 +31,8 @@ class DBPostgreSQLAdapter extends DBSQLAdapter<PostgreSQLExecutionContext>
   static void boot() {
     if (_boot) return;
     _boot = true;
+
+    DBSQLAdapter.boot();
 
     Transaction.registerErrorFilter((e, s) => e is PostgreSQLException);
 
@@ -219,7 +222,7 @@ class DBPostgreSQLAdapter extends DBSQLAdapter<PostgreSQLExecutionContext>
   }
 
   @override
-  Object resolveError(Object error, StackTrace stackTrace) {
+  Object resolveError(Object error, StackTrace stackTrace, Object? operation) {
     if (error is DBPostgreSQLAdapterException) {
       return error;
     } else if (error is PostgreSQLException) {
@@ -238,7 +241,7 @@ class DBPostgreSQLAdapter extends DBSQLAdapter<PostgreSQLExecutionContext>
     }
 
     return DBPostgreSQLAdapterException('error', '$error',
-        parentError: error, parentStackTrace: stackTrace);
+        parentError: error, parentStackTrace: stackTrace, operation: operation);
   }
 
   @override
@@ -925,7 +928,6 @@ class DBPostgreSQLAdapterException extends DBSQLAdapterException {
   String get runtimeTypeNameSafe => 'DBPostgreSQLAdapterException';
 
   DBPostgreSQLAdapterException(String type, String message,
-      {Object? parentError, StackTrace? parentStackTrace})
-      : super(type, message,
-            parentError: parentError, parentStackTrace: parentStackTrace);
+      {super.parentError, super.parentStackTrace, super.operation})
+      : super(type, message);
 }